refactor(error-tracking): move backend.api viewsets to backend.presentation
Relocate the Error tracking DRF viewsets out of \`backend/api/\` and
into the new \`backend/presentation/\` package, in preparation for the
facade architecture rollout.

Pure rename for 12 viewset modules — content is unchanged from master:

- assignment_rules, external_references, fingerprints,
  git_provider_file_link_resolver, grouping_rules, issues, releases,
  spike_detection_config, spike_events, stack_frames, symbol_sets

Plus the matching test file move:

- backend/api/test/test_error_tracking_api.py →
  backend/test/presentation/test_error_tracking_api.py

Internal-import fixups:

- backend/presentation/symbol_sets.py: lazy import of
  ErrorTrackingReleaseSerializer now uses the presentation path.
- backend/hogql_queries/error_tracking_query_runner_v1.py and
  backend/hogql_queries/error_tracking_issue_correlation_query_runner.py:
  ErrorTrackingIssuePreviewSerializer now imported from
  backend.presentation.issues.

Compatibility shim:

- backend/api/__init__.py is rewritten as an explicit shim that
  re-exports each viewset from its new presentation location, so
  external callers (\`posthog/api/__init__.py\`) keep working untouched.
  ErrorTrackingSuppressionRuleViewSet still ships from the old
  backend/api/suppression_rules.py until the new presentation
  infrastructure replaces it in the next PR.

Removed obsolete legacy files:

- backend/api/utils.py
- backend/api/test/test_suppression_rules.py
- backend/api/test/fixtures/source.js.map
