+++ products/error_tracking/backend/presentation/spike_detection_config.py
