+++ products/error_tracking/backend/presentation/stack_frames.py
